requirements.txt
fastapi==0.109.0
pymongo==4.6.1
python-dotenv==1.0.0
uvicorn==0.27.0
motor==3.3.1
